requirements.txt
pandas
numpy
folium
turfpy
earthengine-api
geopandas
geemap 
Pillow 
python-dotenv
ipykernel